working on __sort_objects() method on Scene class in order to return self.__objects sorted in respect to area
 Changes to be committed:
	new file:   TODOstupid.txt
	new file:   main.py
	modified:   pyzui/dynamictileprovider.py
	modified:   pyzui/ferntileprovider.py
	modified:   pyzui/globalmosaictileprovider.py
	modified:   pyzui/magickconverter.py
	modified:   pyzui/mainwindow.py
	modified:   pyzui/mandeltileprovider.py
	modified:   pyzui/mediaobject.py
	modified:   pyzui/osmtileprovider.py
	modified:   pyzui/pdfconverter.py
	modified:   pyzui/physicalobject.py
	modified:   pyzui/ppm.py
	modified:   pyzui/qzui.py
	modified:   pyzui/scene.py
	modified:   pyzui/statictileprovider.py
	modified:   pyzui/stringmediaobject.py
	modified:   pyzui/svgmediaobject.py
	modified:   pyzui/tilecache.py
	modified:   pyzui/tiledmediaobject.py
	modified:   pyzui/tilemanager.py
	modified:   pyzui/tileprovider.py
	modified:   pyzui/tiler.py
	modified:   pyzui/tilestore.py
	modified:   pyzui/webkitconverter.py

+++ pyzui/ppm.py
@@ -18,7 +18,7 @@
 
 """Module for reading PPM images."""
 
-from tiler import Tiler
+from .tiler import Tiler
 
 def read_ppm_header(f):
     """Read the PPM header in the given file object `f` and return a tuple
@@ -44,8 +44,8 @@ def read_ppm_header(f):
         raise IOError("can only load binary PPM (P6 format)")
 
     try:
-        width = long(header[1])
-        height = long(header[2])
+        width = int(header[1])
+        height = int(header[2])
         maxval = int(header[3])
     except ValueError:
         raise IOError("invalid PPM header")
